Fix cancel behaviour for multiple annotations
